# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: Order & Returns Management API
  version: 1.0.0
  description: |
    API specification for ArtiCurated order and return management. Covers order lifecycle,
    returns workflow, and endpoints to inspect state history. Background jobs (invoice/refund)
    are queued by server-side events and are not directly invoked via these endpoints.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /orders:
    post:
      summary: Create a new order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
...
  /orders/{orderId}/returns:
    post:
      summary: Create a return request for a delivered order
      operationId: createReturn
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnCreateRequest'
      responses:
        '201':
          description: Return created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnResponse'
...
components:
  schemas:
    OrderCreateRequest:
      type: object
      required: [customerId, items, totalAmount]
      properties:
        customerId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: double
        externalId:
          type: string
    OrderItem:
      type: object
      required: [sku, name, qty, price]
      properties:
        sku:
          type: string
        name:
          type: string
        qty:
          type: integer
        price:
          type: number
          format: double
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
        customerId:
          type: string
          format: uuid
        totalAmount:
          type: number
          format: double
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ReturnCreateRequest:
      type: object
      required: [requestedBy, refundAmount, reason]
      properties:
        requestedBy:
          type: string
        refundAmount:
          type: number
          format: double
        reason:
          type: string
    ReturnResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        status:
          type: string
        refundAmount:
          type: number
          format: double
        requestedAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
...
  securitySchemes: {}
...